Update url of author page
