--- footer/src/asset-link.js
@@ -1,15 +1,0 @@
-import FBIcon from '../static/fb-logo_hover@2x.svg'
-import IGIcon from '../static/IG-logo_hover.svg'
-import LineIcon from '../static/Line-icon_hover@2x.svg'
-import MediumIcon from '../static/medium-logo_hover@2x.svg'
-import GithubIcon from '../static/github-logo_hover@2x.svg'
-import RSSIcon from '../static/rss-logo_hover@2x.svg'
-
-export default {
-  FBIcon,
-  IGIcon,
-  LineIcon,
-  MediumIcon,
-  GithubIcon,
-  RSSIcon,
-}
