+++ footer/src/styles/theme.js
@@ -1,22 +1,46 @@
-import { pageThemes } from 'shared/configs'
-import { colors } from 'shared/common-variables'
-
-export const selectTextColor = (pageTheme) => {
-  switch (pageTheme) {
-    case pageThemes.dark:
-      return colors.footerTextLight
-    case pageThemes.bright:
-    default:
-      return colors.footerTextDark
-  }
-}
-
-export const selectBgColor = (pageTheme) => {
-  switch (pageTheme) {
-    case pageThemes.dark:
-      return colors.photographyBg
-    case pageThemes.bright:
-    default:
-      return colors.footerBg
-  }
+export const styles = {
+  reporterLogo: {
+    width: {
+      desktop: 258, // px
+      hd: 300, // px
+    },
+    height: {
+      desktop: 45, // px
+      hd: 55, // px
+    },
+  },
+  icon: {
+    width: {
+      tabletAbove: 30, // px
+      mobile: 40, // px
+    },
+    height: {
+      tabletAbove: 30, // px
+      mobile: 40, // px
+    },
+  },
+  footerHeight: {
+    desktop: 309, // px
+  },
+  footerContentPadding: {
+    hd: [60, 60, 41, 60], // px
+    desktop: [60, 50, 40, 50], // px
+    tablet: [50, 35, 0, 35], // px
+    mobile: [60, 40, 60, 35], // px
+  },
+  footerContentMaxWidth: {
+    desktop: 1024, // px
+    hd: 1440, // px
+  },
+  contentRow: {
+    width: {
+      tablet: 531, // px
+      desktop: 682, // px
+      hd: 908, // px
+    },
+    height: {
+      desktop: 123, // px
+      hd: 96, // px
+    },
+  },
 }
